app: prevent file overwrite in extract command
The extract can generate the name of the Crash Log file automatically
when no output path are specified (or when the specified path points to
a directory).
If the generated path already exists in the file system, the extract
command will now try to automatically find another filename that does
not conflict with the files currently on the file system.

If the user specifies a path that already exists, an error is reported
and the existing file won't be overwritten.

+++ app/src/extract.rs
@@ -36,29 +36,44 @@ pub fn extract(output_path: Option<&Path>, sources: Vec<CrashLogSource>) {
         log::error!("{}", Error::NoCrashLogFound);
     }
 
-    for (i, crashlog) in result.iter().enumerate() {
-        let mut path = if let Some(output_path) = output_path {
-            let mut path = output_path.to_path_buf();
-            if output_path.is_dir() {
-                path.push(format!("{}.crashlog", crashlog.metadata))
+    for crashlog in result {
+        let mut idx = 0;
+
+        loop {
+            let filename = if idx > 0 {
+                format!("{}-{}.crashlog", crashlog.metadata, idx)
+            } else {
+                format!("{}.crashlog", crashlog.metadata)
+            };
+
+            let path = if let Some(output_path) = output_path {
+                let mut path = output_path.to_path_buf();
+                if output_path.is_dir() {
+                    path.push(filename);
+                }
+                path
+            } else {
+                PathBuf::from(filename)
+            };
+
+            if path.exists() {
+                if let Some(output_path) = output_path
+                    && output_path.is_file()
+                {
+                    log::error!("{}: file already exists.", path.display());
+                    break;
+                }
+
+                idx += 1;
+                continue;
             }
-            path
-        } else {
-            PathBuf::from(format!("{}.crashlog", crashlog.metadata))
-        };
 
-        if result.len() > 1
-            && let Some(filename) = path.file_stem()
-        {
-            path.set_file_name(format!(
-                "{}-{i}.crashlog",
-                PathBuf::from(filename).display()
-            ))
-        }
+            println!("{}", path.display());
+            if let Err(err) = std::fs::write(path, crashlog.to_bytes()) {
+                log::error!("Failed to write Crash Log file: {err}");
+            }
 
-        println!("{}", path.display());
-        if let Err(err) = std::fs::write(path, crashlog.to_bytes()) {
-            log::error!("Failed to write Crash Log file: {err}");
+            break;
         }
     }
 }
